Crate check_guide

This crate verifies the correctness of every Hermes command in the guide by:

Extracting every line in the guide with ‘{{#template templates/commands/hermes}}’, a macro call for mdbook template.

Replace every template call with the content of the template. It will replace the macro call with what should be a Hermes command.

Check that an EntryPoint can be created from the command.

Structs

EXCLUSIONS 🔒

FILEREADER 🔒

GUIDE_PATH 🔒

TEMPLATES_PATH 🔒

TEMPLATE_RE 🔒

Enums

ParseError 🔒

Functions

check_correctness 🔒

main 🔒

verify_file 🔒

Verifies that every template macro call in the file can be replaced by a valid Hermes command. Returns the number of invalid commands found.

verify_line 🔒

Extracts macro calls from a command, replace them with the content of the template and check that the command is correct. Returns the number of errors encountered.